Reforma

+++ src/main/java/com/proyecto/repository/ClientesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.proyecto.models.Clientes;
 
 @Repository
-public interface ClientesRepository extends CrudRepository<Clientes,Integer>{
+public interface ClientesRepository extends CrudRepository<Clientes,Long>{
 
 }
