+++ src/main/java/com/proyecto/repository/PedidosRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.proyecto.models.Pedidos;
 
 @Repository
-public interface PedidosRepository extends CrudRepository<Pedidos,Integer>{
+public interface PedidosRepository extends CrudRepository<Pedidos,Long>{
 
 }
