+++ src/main/java/com/proyecto/repository/ProductosRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.proyecto.models.Productos;
 
 @Repository
-public interface ProductosRepository extends CrudRepository<Productos,Integer>{
+public interface ProductosRepository extends CrudRepository<Productos,Long>{
 
 }
